fix(circuit_breaker): UpdateAfterTradeClose cast $1 to numeric
Round 5 smoke v10 Scenario 1 暴露: 现 trade close pipeline 成功但
side-effect rollup 报 invalid input syntax for type integer: "-1.368".

根因: pgx 看到 `$1 < 0` (0 是 integer literal) 推断 $1 为 integer,
然后 numeric.Decimal -1.368 文本编码送过去 → PG 试图 cast "-1.368"
到 integer → 22P02 invalid_text_representation.

修: 全部 $1 引用加 ::numeric 显式 cast.

不影响 trade close 主流程 (UpdateTradeClosed/InsertTradeExit 都是
单独 query, 不复用 $1). Round 6 读 consec_losses + daily_pnl
之前必须有这个 fix.

+++ internal/storage/postgres/queries/circuit_breaker.sql
@@ -67,12 +67,12 @@ WHERE id = 1;
 -- Round 6 will read these for trip evaluation.
 UPDATE circuit_breaker_state
 SET consecutive_losses = CASE
-        WHEN $1 < 0 THEN consecutive_losses + 1
+        WHEN ($1::numeric) < 0 THEN consecutive_losses + 1
         ELSE 0
     END,
     daily_pnl = CASE
-        WHEN daily_pnl_date = $2 THEN daily_pnl + $1
-        ELSE $1
+        WHEN daily_pnl_date = $2 THEN daily_pnl + $1::numeric
+        ELSE $1::numeric
     END,
     daily_pnl_date = $2
 WHERE id = 1;
